+++ src/main/scala/io/iohk/ethereum/merklePatriciaTree/IodbDataSource.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.utils
+package io.iohk.ethereum.merklePatriciaTree
 
 import io.iohk.iodb.{ByteArrayWrapper, LSMStore}
